Move libcloud specific methods to separate class.
